+++ lib/io_loader.dart
@@ -12,15 +12,15 @@ import 'package:http/http.dart';
 import 'package:grizzly_io/grizzly_io.dart';
 
 /// Downloads the TSV file from specified [url] and returns the parsed data
-Future<LabeledTsv> requestLabTsv(String url) async {
+Future<LabeledTable> requestLabTsv(String url) async {
   final client = new Client();
   final Response resp = await client.get(url);
-  return new LabeledTsv.parse(resp.body);
+  return parseLabTsv(resp.body);
 }
 
 /// Reads file at specified path [path] as TSV file
-Future<LabeledTsv> readLabTsv(String path, {Encoding encoding: UTF8}) async {
+Future<LabeledTable> readLabTsv(String path, {Encoding encoding: UTF8}) async {
   final File file = new File(path);
   if (!await file.exists()) throw new Exception('File not found!');
-  return new LabeledTsv.parse(await file.readAsString(encoding: encoding));
+  return parseLabTsv(await file.readAsString(encoding: encoding));
 }