+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.4
+
+- Replaced `LabeledTsv` with `parseLabTsv` and `toLabTsv`
+- Added `LabeledTable` and `Table` to fluently morph columns in a table
+
 ## 0.0.3
 
 - Browser and dart:io based loaders to read file formats from different sources